+++ src/utils/index.node.ts
@@ -17,9 +17,10 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
-export * as Enums from './enums'
-export * as Models from './models'
-export * as SPDX from './spdx'
-export * as Spec from './spec'
-export * as Types from './types'
-// do not export the _helpers, they are for internal use only
+export * from './index.common'
+
+// region node-specifics
+
+export * as NpmJs from './npmjs.node'
+
+// endregion node-specifics