+++ HISTORY.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 <!-- add unreleased items here -->
 
+## 8.2.0 -- 2025-06-05
+
 * Added
   * `factories.FromNodePackageJson.makeExternalReferences()` supports "dist" field ([#1247] via [#1246]) 
   * New symbols under `utils.NpmjsUtility` (via [#1246])